[NayNay] File Restructure: Faucet
- restructured faucet to match the new structure
